Finished grabData helper function read msg body for more details
I am assuming the format of csv files fed to this will be the same as
the one provided initially, so I am not putting a parameter check for
the bounds of a line of entry. In addition, if anyone want to optomize
this please feel free because using scanner is not very efficient.